+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './to-ts';
+export * from './to-request';
 export * from './to-md';
 export * from './to-mock';
 export * from './mock-server';