+++ src/to-mock/index.ts
@@ -2,12 +2,14 @@ import { AriesConfig, Plugin } from '../interface';
 import { generateOutputByPlugin } from '../core';
 import SwaggerParserV2 from '../core/parseV2';
 
-const genMock: Plugin.Function = ({ swagger, options }) => {
+export const genMock: Plugin.Function = ({ swagger, options }) => {
   const { paths } = new SwaggerParserV2(swagger, options);
+  const { basePath } = swagger;
   const mockJson: Record<string, Record<string, any>> = {};
   Object.keys(paths).forEach((path) => {
     Object.keys(paths[path]).forEach((method) => {
-      mockJson[`${method.toLocaleUpperCase()} ${path}`] = paths[path][method].mock.responses;
+      mockJson[`${method.toLocaleUpperCase()} ${basePath}${path}`] =
+        paths[path][method].mock.responses;
     });
   });
   return JSON.stringify(mockJson, null, 2);